requirements2.txt
langchain==0.1.16
langchain-community==0.0.32
langchain-groq==0.1.6
faiss-cpu
sentence-transformers
streamlit
python-dotenv
pypdf
huggingface-hub
